added Support for the Samsung (V)acuum (F)luorescent (D)isplay (M)odule 20L203DA5
It is a simple 20x2 Text Display that is controlled via one single
serial line and supports up to 19200 Baud.

Note that the logic level on the line needs to be inverted and amplified
to 5V to make this display working. A simple transistor would do.

To enable the driver #define USE_DISPLAY_VFDM on your custom build.
To override the default standard baudrate (9600)
   #define VFDM_BAUDRATE 19200  // this is the most you can get
Note that non-standard baud rates need jumper to be set on the VFDM.

A pin has to be assigned VFDM_Tx to recognize the display. There is
no probing possible with a "write-only" interface. If the pin is
assigned, the display is assumed to be there.